Refactor placement of predicateCheckResult
